fix: 宿泊施設の登録フォーム付きボタンを修正
